+++ src/presenter/movies-presenter.js
@@ -314,7 +314,6 @@ export default class MoviesPresenter {
 
   destroy() {
     this._clearBoard({ resetRenderedFilmCardsCount: true, resetSortType: true });
-    console.warn('movies board Уничтожен !');
   }
 
 }
